server/requirements.txt
appdirs==1.4.0
dj-database-url==0.4.2
Django==1.10.5
djangorestframework==3.5.3
gunicorn==19.7.1
packaging==16.8
psycopg2==2.7.3.1
pyparsing==2.1.10
six==1.10.0
whitenoise==3.3.0
